Split class UniqueWord into header and cpp file. Added to CMakeLists.txt. Added tst_files directory for easy testing of text files.
