+++ exercises/matthew_cumber/ex01_basics/tst_files/test.txt
@@ -1,0 +1,4 @@
+the the the a b a
+APPLE apple AppLE
+
+ha ha ha ha ha ha
